Refactor: Split test modules between single and multi process directories
