+++ tests/workers/single_process/test_conan_install.py
@@ -1,5 +1,5 @@
 """
-Test the Conan command functionality, in a single process.
+Test the conan install functionality, in a single process.
 
 Note that this is not how cruiz tends to work, and should be using multiple processes
 to isolate the Conan commands, but this test shows it still works without that
@@ -23,27 +23,8 @@ from cruizlib.interop.message import (
 # pylint: disable=wrong-import-order
 import pytest
 
-# pylint: disable=import-error
-import testexceptions
-
 
 LOGGER = logging.getLogger(__name__)
-
-
-def test_expected_failure(
-    reply_queue_fixture: typing.Tuple[
-        queue.Queue[Message], typing.List[Message], threading.Thread
-    ],
-) -> None:
-    """Test: running conan install incorrect setup, so has an expected failure."""
-    worker = workers_api.install.invoke
-    params = CommandParameters("install", worker)
-    reply_queue, _, watcher_thread = reply_queue_fixture
-    # abusing the type system, as the API used for queue.Queue is the same
-    # as for multiprocessing.Queue
-    worker(reply_queue, params)  # type: ignore[arg-type]
-    with pytest.raises(testexceptions.FailedMessageTestError):
-        watcher_thread.join()
 
 
 @pytest.mark.parametrize(
